legalassistant/pyproject.toml
[project]
name = "legalassistant"
version = "0.1.0"
description = "LegalAssistant using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "python-docx>=1.1.0"
]

[project.scripts]
run_crew = "legalassistant.main:run"
train = "legalassistant.main:train"
replay = "legalassistant.main:replay"
test = "legalassistant.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
